% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.R
\name{acc_pivot_wider}
\alias{acc_pivot_wider}
\title{Pivot Ornitela accelerometry data to long format (xyzxyzxyz...)}
\usage{
acc_pivot_wider(
  data,
  burstMinApart = 6,
  pointsPerBurst = 30,
  removeIncompleteBursts = TRUE
)
}
\arguments{
\item{data}{A data frame of an Ornitela .csv for acc data ("SENSORS").}

\item{burstMinApart}{The frequency of ACC bursts in minutes.}

\item{pointsPerBurst}{The number of points in a burst, a.k.a. hertz * seconds.}

\item{removeIncompleteBursts}{Should incomplete bursts (bursts with fewer
than \code{pointsPerBurst}) be removed from the final data frame? The default
is \code{TRUE}.}
}
\value{
A data frame.
}
\description{
This function takes a data frame made from a .csv downloaded
directly from the Ornitela website, and pivots it into the format
"xyzxyzxyz...", which is used for analyses in the package \code{rabc} on
Github.
}
\examples{
NA
}
