% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerange.R
\name{home_range}
\alias{home_range}
\title{Calculate a home range from a \code{Move} object}
\usage{
home_range(data, method = "AKDE", ...)
}
\arguments{
\item{data}{A \code{Move*} object, or a list of those objects.}

\item{method}{A home range calculation method:
\itemize{
\item \code{"AKDE"}: Autocorrelated kernel density estimate (default), using the
\code{ctmm} package.
\item \code{"dBBMM"}: Dynamic Brownian Bridge Movement Model from the \code{move}
package.
\item \code{"MCP"}: Minimum convex polygon from the \code{adehabitatHR} package.
\item \code{"KDE"}: Kernel density estimate from the \code{adehabitatHR} package.
}}

\item{...}{Additional parameters to pass to the home range function of your
choice.}
}
\value{
A home range object.
}
\description{
This function currently defaults to an autocorrelated kernel
density estimate, under the assumption that your relocation data is very frequent
(1 hour or less, generally). Several other methods are currently supported,
and more will be added.
}
\examples{
NA
}
