% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerange.R
\name{home_range_area}
\alias{home_range_area}
\title{Retrieve the home range size in hectares}
\usage{
home_range_area(data, size = c(95, 50), method = "AKDE")
}
\arguments{
\item{data}{The output from \code{waterfowl::home_range}.}

\item{size}{The percentage of the home range to calculate. Default is both
95\% (home range) and 50\% (core home range).}

\item{method}{The method that produced the data from \code{waterfowl::home_range}:
\code{"AKDE"}, \code{"dBBMM"}, \code{"MCP"}, \code{"KDE"}.}
}
\value{
A data frame with the home range sizes in hectares.
}
\description{
Retrieve the home range size in hectares
}
\examples{
NA

}
