% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcover.R
\name{landcover_percent}
\alias{landcover_percent}
\title{Calculate land cover class percentages for a polygon}
\usage{
landcover_percent(
  data,
  raster.source = "FedData",
  buffer = 0,
  collapse = FALSE,
  pivot_wide = TRUE
)
}
\arguments{
\item{data}{Spatial data of class \code{Move*} or \code{SpatialPointsDataFrame},
or a list of those classes.}

\item{raster.source}{A raster. By default, downloads the National Land Cover
Database using the \code{FedData} package. Otherwise, accepts any raster that can
work with \code{terra::rast()}.}

\item{buffer}{An optional buffer to add to your spatial data, in meters.}

\item{collapse}{If TRUE and the input was a list, collapses the list into a
single data frame.}

\item{pivot_wide}{If TRUE, pivots the resulting data frame so that each value
type (by default land cover classes) is in its own column. Useful if you intend
to model habitat selection.}
}
\value{
A data frame of percentages.
}
\description{
Calculate land cover class percentages for a polygon
}
\examples{
NA
}
