% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcover.R
\name{landcover_points}
\alias{landcover_points}
\title{Determine land cover class at specified point locations}
\usage{
landcover_points(points, raster.source = "FedData", df = FALSE)
}
\arguments{
\item{points}{Spatial data of class \code{Move*}, \code{SpatialPointsDataFrame},
or \code{SpatVector}, or a list of those classes.}

\item{raster.source}{A raster. By default, downloads the National Land Cover
Database using the \code{FedData} package. Otherwise, accepts any raster that can
work with \code{terra::rast()}.}

\item{df}{Should the output be a data frame? If FALSE, returns the input object
with an additional data column "landcover".}
}
\value{
The original data with an added column of land cover types, or a data
frame of the land cover types if df = TRUE.
}
\description{
Determine land cover class at specified point locations
}
\examples{
NA

}
