% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{subsetMove}
\alias{subsetMove}
\title{Subset or split a Move* object}
\usage{
subsetMove(
  data,
  birds = NA,
  date = NA,
  month = NA,
  year = NA,
  DayNight = "no",
  timeSplit = "none"
)
}
\arguments{
\item{data}{A \code{Move*} object.}

\item{birds}{Temporary.}

\item{date}{Temporary.}

\item{month}{Integers 1-12, representing the month of the year.}

\item{year}{Four digit year.}

\item{DayNight}{Should the \code{Move*} object be subset to just daytime or
just nighttime points? Choose \code{day} for day and \code{night} for night.}

\item{timeSplit}{Either \code{daily}, \code{weekly}, \code{monthly}, or \code{yearly}.
Creates multiple \code{Move} objects at the given frequency. For example, if
your data consists of two full months and you choose \code{timesplit = "daily"},
the function will return a list of 60+ \code{Move} objects, one for each day.}
}
\value{
A \code{Move} object or a list of \code{Move} objects.
}
\description{
A function to simplify subsetting or splitting up a \code{Move}
object. Allows for pulling out a single month and/or year, pulling out day
vs. night, and/or splitting into smaller \code{Move} objects for each day,
week, month, or year.
}
\examples{
NA

}
