% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{time_of_day}
\alias{time_of_day}
\title{Find the time of day}
\usage{
time_of_day(move, crepuscular = FALSE)
}
\arguments{
\item{move}{A \code{Move*} object or a vector of \code{POSIXct} times.}

\item{crepuscular}{Should the time of day include crepuscular categories
(dawn & dusk)? Default is FALSE.}
}
\value{
A vector of strings for the time of day.
}
\description{
A simple wrapper around \code{suncalc::getSunlightTimes}, modified
from \code{amt::time_of_day}. Finds the category of time of day (day vs. night)
given time stamps.
}
\examples{
NA
}
