% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfowl-package.R
\docType{package}
\name{waterfowl-package}
\alias{waterfowl}
\alias{waterfowl-package}
\title{waterfowl: Functions for Waterfowl Research}
\description{
A group of wrapper functions to simplify and streamline waterfowl research, primarily by graduate students and state biologists. At present, this package focuses on spatial and tracking data and analsyses, but hopes to expand to all aspects of waterfowl research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/weberbirding/waterfowl}
  \item Report bugs at \url{https://github.com/weberbirding/waterfowl/issues}
}

}
\author{
\strong{Maintainer}: David Weber \email{dweber1594@gmail.com}

}
\keyword{internal}
