% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerange.R
\name{write_home_range}
\alias{write_home_range}
\title{Write a home range to a shapefile}
\usage{
write_home_range(data, method = "AKDE")
}
\arguments{
\item{data}{The output of \code{waterfowl::home_range}.}

\item{method}{The method that produced your home range from
\code{waterfowl::home_range}.}
}
\value{
Write a home range shapefile to your current working direcotry.
}
\description{
Write a home range to a shapefile
}
\examples{
NA

}
